[Uptime] allow an administrator to enable and disable monitor management (#128223)
* uptime - allow an administrator to enable monitor management

+++ x-pack/plugins/uptime/server/lib/synthetics_service/synthetics_service.ts
@@ -191,32 +191,25 @@ export class SyntheticsService {
     }
   }
 
-  async getOutput(request?: KibanaRequest) {
-    if (!this.apiKey) {
-      try {
-        this.apiKey = await getAPIKeyForSyntheticsService({ server: this.server, request });
-      } catch (err) {
-        this.logger.error(err);
-        throw err;
-      }
+  async getApiKey() {
+    try {
+      this.apiKey = await getAPIKeyForSyntheticsService({ server: this.server });
+    } catch (err) {
+      this.logger.error(err);
+      throw err;
     }
 
-    if (!this.apiKey) {
-      const error = new APIKeyMissingError();
-      this.logger.error(error);
-      throw error;
-    }
+    return this.apiKey;
+  }
 
-    this.logger.debug('Found api key and esHosts for service.');
-
+  async getOutput(apiKey: SyntheticsServiceApiKey) {
     return {
       hosts: this.esHosts,
-      api_key: `${this.apiKey.id}:${this.apiKey.apiKey}`,
+      api_key: `${apiKey?.id}:${apiKey?.apiKey}`,
     };
   }
 
   async pushConfigs(
-    request?: KibanaRequest,
     configs?: Array<
       SyntheticsMonitorWithId & {
         fields_under_root?: boolean;
@@ -229,9 +222,16 @@ export class SyntheticsService {
       this.logger.debug('No monitor found which can be pushed to service.');
       return;
     }
+
+    this.apiKey = await this.getApiKey();
+
+    if (!this.apiKey) {
+      return null;
+    }
+
     const data = {
       monitors,
-      output: await this.getOutput(request),
+      output: await this.getOutput(this.apiKey),
     };
 
     this.logger.debug(`${monitors.length} monitors will be pushed to synthetics service.`);
@@ -245,7 +245,6 @@ export class SyntheticsService {
   }
 
   async runOnceConfigs(
-    request?: KibanaRequest,
     configs?: Array<
       SyntheticsMonitorWithId & {
         fields_under_root?: boolean;
@@ -257,9 +256,15 @@ export class SyntheticsService {
     if (monitors.length === 0) {
       return;
     }
+    this.apiKey = await this.getApiKey();
+
+    if (!this.apiKey) {
+      return null;
+    }
+
     const data = {
       monitors,
-      output: await this.getOutput(request),
+      output: await this.getOutput(this.apiKey),
     };
 
     try {
@@ -283,9 +288,16 @@ export class SyntheticsService {
     if (monitors.length === 0) {
       return;
     }
+
+    this.apiKey = await this.getApiKey();
+
+    if (!this.apiKey) {
+      return null;
+    }
+
     const data = {
       monitors,
-      output: await this.getOutput(request),
+      output: await this.getOutput(this.apiKey),
     };
 
     try {
@@ -296,12 +308,23 @@ export class SyntheticsService {
     }
   }
 
-  async deleteConfigs(request: KibanaRequest, configs: SyntheticsMonitorWithId[]) {
+  async deleteConfigs(configs: SyntheticsMonitorWithId[]) {
+    this.apiKey = await this.getApiKey();
+
+    if (!this.apiKey) {
+      return null;
+    }
+
     const data = {
       monitors: this.formatConfigs(configs),
-      output: await this.getOutput(request),
+      output: await this.getOutput(this.apiKey),
     };
     return await this.apiClient.delete(data);
+  }
+
+  async deleteAllConfigs() {
+    const configs = await this.getMonitorConfigs();
+    return await this.deleteConfigs(configs);
   }
 
   async getMonitorConfigs() {
@@ -359,14 +382,6 @@ export class SyntheticsService {
     return configs.map((config: Partial<MonitorFields>) =>
       formatMonitorConfig(Object.keys(config) as ConfigKey[], config)
     );
-  }
-}
-
-class APIKeyMissingError extends Error {
-  constructor() {
-    super();
-    this.message = 'API key is needed for synthetics service.';
-    this.name = 'APIKeyMissingError';
   }
 }
